Refactoring
Fixed SRP violation.
Complete redesign of all facilities and systems in favor of partial ECS.
The Render() method has been removed from all objects and the scene.
CameraObject has been removed and replaced with a Camera2D/Camera3D component
Shaders are no longer scattered throughout the code, but are stored only in the RenderSystem.
Application places less emphasis on RenderWindow, with some functionality moved to RenderSystem
A UI with basic functionality has been added to the game.
MSAA has been added back to the game, but as a separate class.
The player's movement (as well as other objects) is now controlled by the MovementSystem.
Collider no longer checks for collisions; this is now handled by the PhysicsSystem.
The EPA, GJK, and RayCast algorithms are now separate classes and not part of Collider.
RigidBody no longer handles collisions; this is now handled by the PhysicsSystem.
Added temporary MatrixPlus.h, part of the game will be ported to Neutronium in the future.
TargetingSystem and BehaviorSystem no longer work with the scene directly, they now need to register objects with themselves.
Added QueryBus, which is similar to Request-Response, but is not intended for the server.